Backend/facebook_arch/requirements.txt
langchain>=0.1.16,<0.2.0
langchain-openai>=0.1.3,<0.2.0
neo4j>=5.18.0,<6.0.0
python-dotenv>=1.0.1,<2.0.0
langchain-core>=0.1.41,<0.2.0
langchain-community>=0.0.32,<0.1.0
chainlit>=1.0.0
streamlit>=1.32.0,<2.0.0
pandas>=2.0.0,<3.0.0
nest-asyncio>=1.5.0
openai>=1.14.0,<2.0.0
asyncio

plotly>=5.0.0,<6.0.0

langchain-text-splitters==0.2.2
langchain-community==0.2.7
mem0ai

